This is a new direction for the password manager.
I'm leaving the old files up for my reference but the new file named password_manager.py is a standalone without a GUI. It is all done in a command line.

I made a new password manager for my personal use. It's a simple Python script that uses the cryptography library to encrypt and decrypt a JSON file containing a list of websites and their associated passwords. The passwords are generated using the Fernet symmetric encryption algorithm, which is a modern and secure encryption method that is also fast and efficient.

The script also includes a function to generate a random password of a specified length. This is useful for creating new passwords for websites that don't have one already.

To use the script, you'll need to have Python installed on your computer. Once you have Python installed, you can run the script by opening a terminal or command prompt and navigating to the directory where the script is located. Then, you can simply run the script by typing `python password_manager.py` and pressing Enter.

When you run the script for the first time, it will prompt you to enter a master password. This password will be used to encrypt and decrypt the JSON file containing the websites and passwords. It's important to choose a strong and secure password, as this will be the only way to access your passwords.

Once you've entered the master password, you can use the script to add new websites and generate passwords for them.